fixed non integer indices errors

+++ fourier_tools.py
@@ -21,13 +21,13 @@ def stft(sig, numPoints,frameSize, overlapFac=0.5, window=np.hanning):
     hopSize = int(frameSize - np.floor(overlapFac * frameSize))
     
     # zeros at beginning (thus center of 1st window should be for sample nr. 0)
-    samples = np.append(np.zeros(np.floor(frameSize/2.0)), sig)    
+    samples = np.append(np.zeros(np.int(np.floor(frameSize/2.0))), sig)    
     # cols for windowing
     cols = np.ceil( (len(samples) - frameSize) / float(hopSize)) + 1
     # zeros at end (thus samples can be fully covered by frames)
     samples = np.append(samples, np.zeros(frameSize))
     
-    frames = stride_tricks.as_strided(samples, shape=(cols, frameSize), strides=(samples.strides[0]*hopSize, samples.strides[0])).copy()
+    frames = stride_tricks.as_strided(samples, shape=(np.int(cols), frameSize), strides=(samples.strides[0]*hopSize, samples.strides[0])).copy()
     frames *= win
 
     return np.fft.rfft(frames,n=numPoints)   
@@ -74,18 +74,18 @@ def logscale_spec(spec, sr=44100, factor=20.):
     newspec = np.complex128(np.zeros([timebins, len(scale)]))
     for i in range(0, len(scale)):
         if i == len(scale)-1:
-            newspec[:,i] = np.sum(spec[:,scale[i]:], axis=1)
+            newspec[:,i] = np.sum(spec[:,np.int(scale[i]):], axis=1)
         else:        
-            newspec[:,i] = np.sum(spec[:,scale[i]:scale[i+1]], axis=1)
+            newspec[:,i] = np.sum(spec[:,np.int(scale[i]):np.int(scale[i+1])], axis=1)
     
     # list center freq of bins
     allfreqs = np.abs(np.fft.fftfreq(freqbins*2, 1./sr)[:freqbins+1])
     freqs = []
     for i in range(0, len(scale)):
         if i == len(scale)-1:
-            freqs += [np.mean(allfreqs[scale[i]:])]
+            freqs += [np.mean(allfreqs[np.int(scale[i]):])]
         else:
-            freqs += [np.mean(allfreqs[scale[i]:scale[i+1]])]
+            freqs += [np.mean(allfreqs[np.int(scale[i]):np.int(scale[i+1])])]
     
     return newspec, freqs
